backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
scikit-learn
pandas
openpyxl
pydantic[email]
email-validator
bcrypt>=4.0.0
websockets>=10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Performance optimization packages (easy to uninstall: pip uninstall numba llvmlite orjson cython)
numba>=0.58.0
orjson>=3.9.0
cython>=3.0.0
